Add explicit cast to double to fix compile time errors about ambiguous overloaded functions.
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@91 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ common/math.hpp
@@ -33,7 +33,7 @@ static inline uint32_t toPow2(uint32_t num) {
 vector<int> closest_point_on_line(int x1, int y1, int x2, int y2, int x3, int y3);
 
 static inline double dist_between_points(int x1, int y1, int x2, int y2) {
-	return sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
+	return sqrt(double((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1)));
 }
 
 #endif
